Refactor the main structure and place dependency management into the hands of Conan
* Started refactoring

* Test

* Added second Workflow

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Fixed tests

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test wheels

* Test

* Test

* Test

* Test

* Test

* Test Warnings

* Test

* Test Cmake 3.8 and find python

* Test

* Test

* Test

* Test

* Test

* Test

* Fixed conan override outputs

* Fixed python module generation

* Test

* Test

* Cleaned workflows and cmakelist

* Disable build docs by default

* Added resources to python module + Updated python docs

* Updated GameRenderer section + Updated mains loading correct default config

* Updated installation docs

* Disabled workflow wheels on push

* Updated docs references  + yamls

* Added assets

* Removed old assets

+++ src/stratega/src/GUI/GameRenderer.cpp
@@ -1,12 +1,12 @@
 #include <cassert>
 
-#include <stratega/GUI/GameRenderer.h>
-#include <stratega/GUI/RTSGameRenderer.h>
-#include <stratega/GUI/TBSGameRenderer.h>
+#include <Stratega/GUI/GameRenderer.h>
+#include <Stratega/GUI/RTSGameRenderer.h>
+#include <Stratega/GUI/TBSGameRenderer.h>
 
 namespace SGA
 {
-	std::unique_ptr<GameRenderer> createRenderer(const GameType& type, SGA::Vector2f& resolution)
+	std::unique_ptr<GameRenderer> createRenderer(const GameType& type, SGA::Vector2i& resolution)
 	{
 		if (type == GameType::TBS)
 		{
